EDA Deployment/Usman Ayaz/requirements.txt
streamlit==1.1.0
pandas==1.3.4
matplotlib==3.4.3
seaborn==0.11.2
pillow == 8.4.0
